settings save between uses. range and checkboxes limit crimes displayed completely. fixed all recognized bugs need to fix radius bar. doesnt save the physical changes of the bar.

+++ app/src/main/java/com/example/akiva/crimestoppers/Parser.java
@@ -100,7 +100,7 @@ public class Parser {
 		}
 	}
 
-	public TreeSet<Crime> nearestCrimes(double lat, double lon, int radius) {
+	public TreeSet<Crime> nearestCrimes(double lat, double lon, double radius) {
 
 		TreeSet<Crime> tree = new TreeSet<Crime>();
 		for (Crime c : data) {
